Moved visitor count
Trying a new location for the visitor count to see if the CSS doesn't screw it up this time. The preview for this is honestly atrocious, when I removed the GitHub header image it would still show up in the preview so I had no idea what things really looked like.