Sped up cron job
I want it to update my hours spent programming more often. So far several hours and projects are unaccounted for, and I'm hoping this will fix that. If not I need to create an issue on the waka repository, because I can see the hours and projects on the official WakaTime dashboard.